modify : id 필드 타입 변경

+++ src/main/java/com/risingtest/wanted/src/user/UserProvider.java
@@ -81,7 +81,7 @@ public class UserProvider {
         }
 
         if(user.getPassword().equals(encryptPwd)){
-            int userIdx = user.getId();
+            long userIdx = user.getId();
             String jwt = jwtService.createJwt(userIdx);
             return new PostLoginRes(userIdx,jwt);
         }
